src/requirements.txt
cloudshell-shell-core
tornado
cloudshell-automation-api
